+++ src/main/java/net/mohron/skyclaims/command/team/CommandKick.java
@@ -79,7 +79,7 @@ public class CommandKick extends CommandBase.IslandCommand {
         PrivilegeType type = island.getPrivilegeType(user);
         user.getPlayer().ifPresent(p -> {
             if (island.getPlayers().contains(p)) {
-                p.setLocationSafely(island.getWorld().getSpawnLocation());
+                p.setLocationSafely(PLUGIN.getConfig().getWorldConfig().getSpawn());
             }
             p.sendMessage(Text.of(TextColors.RED, "You have been removed from ", island.getName(), TextColors.RED, "!"));
         });